Conectar Router vue y crear bases de registro, login y de editar usuario, ademas de home
