+++ src/main.js
@@ -1,7 +1,7 @@
 import { createApp } from 'vue';
 import App from './App.vue';
 import { createClient } from "@supabase/supabase-js";
-
+import router from './router'
 const supabaseUrl = 'https://jaubayxbwflblrthehdj.supabase.co/';
 const supabaseKey = 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZSIsInJlZiI6ImphdWJheXhid2ZsYmxydGhlaGRqIiwicm9sZSI6ImFub24iLCJpYXQiOjE3MTI0MzM1MDIsImV4cCI6MjAyODAwOTUwMn0.tGggr751oq5X9NZ3PVGcv5jnBiQOicsbBZPViOHMtYI';
 const supabase = createClient(supabaseUrl, supabaseKey);
@@ -10,4 +10,6 @@ const app = createApp(App);
 
 app.config.globalProperties.$supabase = supabase;
 
-app.mount('#app');
+createApp(App)
+  .use(router)
+  .mount('#app')
