+++ src/style.css
@@ -77,3 +77,18 @@ button:focus-visible {
     background-color: #f9f9f9;
   }
 }
+
+nav, main {
+  border: 2px solid #000;
+  margin-bottom: 10px;
+  padding: 10px;
+}
+
+nav > a + a {
+  margin-left: 10px;
+}
+
+h2 {
+  border-bottom: 1px solid #ccc;
+  margin: 0 0 20px;
+}
